middle version

+++ src/main/java/DBConnection.java
@@ -3,29 +3,33 @@ import java.sql.*;
 public class DBConnection
 {
     private static Connection connection;
-    private static StringBuilder builder = new StringBuilder();
-
+   // private static StringBuilder builder = new StringBuilder();
+    private static boolean exit = false;
     private static String dbName = "learn";
     private static String dbUser = "root";
     private static String dbPass = "23019088";
 
     public static Connection getConnection()
     {
-        if(connection == null)
+        if(connection == null )
         {
             try {
                 connection = DriverManager.getConnection(
                     "jdbc:mysql://localhost:3306/" + dbName +
                     "?useUnicode=true&useJDBCCompliantTimezoneShift=true&useLegacyDatetimeCode=false&serverTimezone=UTC" +
                              "&user=" + dbUser + "&password=" + dbPass);
-                connection.createStatement().execute("DROP TABLE IF EXISTS voter_count");
-                connection.createStatement().execute("CREATE TABLE voter_count(" +
-                        "id INT NOT NULL AUTO_INCREMENT, " +
-                        "name VARCHAR(50) NOT NULL, " +
-                        "birthDate DATE NOT NULL, " +
-                        "count INT NOT NULL, " +
-                        "PRIMARY KEY(id)," +
-                        "UNIQUE KEY name_date(birthDate,name))");
+
+                if(!exit) {
+                    connection.createStatement().execute("DROP TABLE IF EXISTS voter_count");
+                    connection.createStatement().execute("CREATE TABLE voter_count(" +
+                            "id INT NOT NULL AUTO_INCREMENT, " +
+                            "name VARCHAR(50) NOT NULL, " +
+                            "birthDate DATE NOT NULL, " +
+                            "count INT NOT NULL, " +
+                            "PRIMARY KEY(id)," +
+                            "UNIQUE KEY name_date(birthDate,name))");
+                    exit = true;
+                }
             } catch (SQLException e) {
                 e.printStackTrace();
             }
@@ -33,17 +37,19 @@ public class DBConnection
         return connection;
     }
 
-    public static void executeMultiInsert() throws SQLException {
+    public static void executeMultiInsert(StringBuilder builder) throws SQLException {
         String sql = "INSERT INTO voter_count(name, birthDate, count) " +
                 "VALUES" + builder.toString() +
                 "ON DUPLICATE KEY UPDATE count = values(count) + 1";
-        DBConnection.getConnection().createStatement().execute(sql);
+        synchronized (DBConnection.class) {
+            DBConnection.getConnection().createStatement().execute(sql);
+        }
     }
 
-    public static void countVoter(String name, String birthDay) throws SQLException
+    /*public synchronized static void countVoter(String name, String birthDay) throws SQLException
     {
         builder.append((builder.length()==0 ? "" : ",") +
-                "('" + name + "', '" + birthDay + "',1)");
+                "('" + name + "', '" + birthDay + "',1)");*/
 
 
 
@@ -66,7 +72,7 @@ public class DBConnection
                     .execute("UPDATE voter_count SET `count`=`count`+1 WHERE id=" + id);
         }
         rs.close();*/
-    }
+
 
     public static void printVoterCounts() throws SQLException
     {
